+++ src/App.sass
@@ -1,5 +1,34 @@
 // SASS file to customize Bulma -- Update variables and import bulma components (or whole file)
-// I removed the .numbers and .hightlight class from bulma/sass/components/others.sass 
-@charset "utf-8";
 
-@import "../node_modules/bulma/bulma.sass";
+// --changes to resolve class name pollution--
+// # commented out .numbers and .hightlight class from bulma/sass/components/others.sass 
+// # changed .tag class to .butag 
+
+// *****Import Customized Modules from src/assets/css ******
+// (edited .sass files are included in source, not installed @ buildtime)
+
+@charset "utf-8"
+
+// @import "../node_modules/bulma/bulma.sass"; 
+@import "../node_modules/bulma/sass/utilities/_all"
+@import "../node_modules/bulma/sass/base/_all"
+// @import "../node_modules/bulma/sass/elements/_all"
+@import "../node_modules/bulma/sass/elements/box"
+@import "../node_modules/bulma/sass/elements/button"
+@import "../node_modules/bulma/sass/elements/container"
+@import "../node_modules/bulma/sass/elements/content"
+@import "../node_modules/bulma/sass/elements/icon"
+@import "../node_modules/bulma/sass/elements/image"
+@import "../node_modules/bulma/sass/elements/notification"
+@import "../node_modules/bulma/sass/elements/progress"
+@import "../node_modules/bulma/sass/elements/table"
+// @import "../node_modules/bulma/sass/elements/tag"
+@import "./assets/sass/tag.sass"
+@import "../node_modules/bulma/sass/elements/title"
+// @import "../node_modules/bulma/sass/elements/other"
+@import "./assets/sass/other.sass"
+@import "../node_modules/bulma/sass/form/_all"
+@import "../node_modules/bulma/sass/components/_all"
+@import "../node_modules/bulma/sass/grid/_all"
+@import "../node_modules/bulma/sass/helpers/_all"
+@import "../node_modules/bulma/sass/layout/_all"
