+++ src/assets/sass/utilities/_all.sass
@@ -1,0 +1,9 @@
+/* Bulma Utilities */
+@charset "utf-8"
+
+@import "initial-variables"
+@import "functions"
+@import "derived-variables"
+@import "mixins"
+@import "controls"
+@import "extends"
