+++ src/assets/sass/utilities/animations.sass
@@ -1,0 +1,1 @@
+@warn "The animations.sass file has MOVED. It is now in the /base folder. Please import sass/base/animations instead."
